Allow running in Dockerfile
I was unable to install lucaschessR dependencies on my host system, as a precise Python version is rquired.

Add a Dockerfile that can be used to start LucasChessR reproducibly, no matter the Python version and other details on the host, as long as it's running X11.

Also make `bin/LucasR.py` a proper executable.